gcc12 fixes, LFNST and Intra CU speedups, cleanups, minor improvements

+++ source/Lib/CommonLib/Buffer.cpp
@@ -228,9 +228,9 @@ void fillN_CuCore( CodingUnit** ptr, ptrdiff_t ptrStride, int width, int height,
 }
 
 void sampleRateConvCore( const std::pair<int, int> scalingRatio, const std::pair<int, int> compScale,
-                         const Pel* orgSrc, const int orgStride, const int orgWidth, const int orgHeight,
+                         const Pel* orgSrc, const ptrdiff_t orgStride, const int orgWidth, const int orgHeight,
                          const int beforeScaleLeftOffset, const int beforeScaleTopOffset,
-                         Pel* scaledSrc, const int scaledStride, const int scaledWidth, const int scaledHeight,
+                         Pel* scaledSrc, const ptrdiff_t scaledStride, const int scaledWidth, const int scaledHeight,
                          const int afterScaleLeftOffset, const int afterScaleTopOffset,
                          const int bitDepth, const bool useLumaFilter,
                          const bool horCollocatedPositionFlag, const bool verCollocatedPositionFlag )
